Refactor imports of campaigns.py
